requirements.txt
pytorch>=1.6.0, <=1.10.0  

opencv-python
easydict>=1.10
tqdm>=4.41.1
pyyaml>=6.0
scipy
matplotlib
tensorboard

grpcio  # for hai
grpcio-tools  # for hai
requests  # for hai dataset_utils
